refactor: Rename `lib/plugins/package/lib/packageService.js`
