+++ lib/plugins/package/lib/package-service.js
