+++ lib/plugins/package/package.js
@@ -5,7 +5,7 @@ const path = require('path');
 const ServerlessError = require('../../serverless-error');
 const cliCommandsSchema = require('../../cli/commands-schema');
 const zipService = require('./lib/zipService');
-const packageService = require('./lib/packageService');
+const packageService = require('./lib/package-service');
 
 class Package {
   constructor(serverless, options) {
