+++ test/unit/lib/plugins/package/lib/package-service.test.js
